requirements.txt
requests==2.31.0
kafka-python==2.0.2
python-dotenv==1.0.0
python-dateutil==2.8.2
pytest==8.0.0 